requirements.txt
bcrypt
fastapi
uvicorn
motor
pydantic
email-validator
python-dotenv
passlib
passlib[bcrypt]
pymongo
